refactor(rust): move system and message to ockam
