+++ implementations/rust/Cargo.toml
@@ -5,13 +5,11 @@ members = [
     "daemon",
     "examples",
     "kex",
-    "message",
     "vault/traits",
     "vault/software",
     "vault/file",
     "vault/ffi",
     "router",
-    "system",
     "queue_topic",
     "transport",
     "node",
@@ -26,13 +24,11 @@ default-members = [
     "daemon",
     "examples",
     "kex",
-    "message",
     "vault/traits",
     "vault/software",
     "vault/file",
     "vault/ffi",
     "router",
-    "system",
     "queue_topic",
     "transport",
     "node",
