+++ implementations/rust/daemon/src/node.rs
@@ -12,14 +12,14 @@ use crate::source::StdinWorker;
 //     ockam_daemon::initiator::StdinWorker
 // }
 
+use ockam::message::{Address, RouterAddress};
 use ockam::secure_channel::*;
+use ockam::system::commands::{OckamCommand, WorkerCommand};
 use ockam_kex::{
     xx::{XXInitiator, XXNewKeyExchanger, XXResponder},
     CipherSuite,
 };
-use ockam_message::message::{Address, RouterAddress};
 use ockam_router::router::Router;
-use ockam_system::commands::{OckamCommand, WorkerCommand};
 use ockam_transport::tcp::TcpManager;
 use ockam_vault_file::ockam_vault::types::*;
 use ockam_vault_file::ockam_vault::*;
