--- implementations/rust/message/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "ockam-message"
-version = "0.1.0"
-authors = ["Ockam Developers"]
-edition = "2018"
-
-[lib]
-crate-type = ["rlib"]
-
-[profile.release]
-lto = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-hex = "0.4.2"
