+++ implementations/rust/node/Cargo.toml
@@ -19,12 +19,10 @@ crate-type = ["rlib"]
 ockam = { "version" = "0.1", path = "../ockam" }
 ockam-router = { version = "0.1", path = "../router" }
 ockam-transport = { version = "0.1", path = "../transport"}
-ockam-message = { version = "0.1", path = "../message"}
 ockam-common = { version = "0.1", path = "../common"}
 ockam-kex = { version = "0.1", path = "../kex" }
 ockam-vault-software = { version = "0.1", path = "../vault/software" }
 ockam-worker = { version = "0.1", path = "../worker"}
-ockam-system = { version = "0.1", path = "../system" }
 
 hex = "0.4.2"
 structopt = { version = "0.3.20", default-features = false }
