+++ implementations/rust/node/src/blast_worker.rs
@@ -1,8 +1,7 @@
+use ockam::message::{Address, AddressType, Message, MessageType, Route, RouterAddress};
 use ockam::secure_channel::CHANNEL_ZERO;
-use ockam_message::message::{Address, AddressType, Message, MessageType, Route, RouterAddress};
-use ockam_system::commands::OckamCommand::{Router, Worker};
-use ockam_system::commands::{OckamCommand, RouterCommand, WorkerCommand};
-use std::str;
+use ockam::system::commands::OckamCommand::{Router, Worker};
+use ockam::system::commands::{OckamCommand, RouterCommand, WorkerCommand};
 
 pub struct BlastWorker {
     rx: std::sync::mpsc::Receiver<OckamCommand>,
