+++ implementations/rust/node/src/hello_worker.rs
@@ -1,6 +1,6 @@
+use ockam::message::{Address, AddressType, Message, MessageType, Route, RouterAddress};
 use ockam::secure_channel::CHANNEL_ZERO;
-use ockam_message::message::{Address, AddressType, Message, MessageType, Route, RouterAddress};
-use ockam_system::commands::{OckamCommand, RouterCommand, WorkerCommand};
+use ockam::system::commands::{OckamCommand, RouterCommand, WorkerCommand};
 use std::str;
 
 pub struct HelloWorker {
