+++ implementations/rust/ockam/src/lib.rs
@@ -1,12 +1,9 @@
 #[macro_use]
-pub extern crate ockam_message as message;
-
+pub mod message;
 pub mod secure_channel;
+pub mod system;
+
 pub use ockam_common as common;
 pub use ockam_kex as kex;
 pub use ockam_queue_topic as queue_topic;
-pub use ockam_router as router;
-pub use ockam_system as system;
-pub use ockam_transport as transport;
 pub use ockam_vault as vault;
-pub use ockam_worker as worker;
