+++ implementations/rust/ockam/src/secure_channel/error.rs
@@ -1,6 +1,6 @@
+use crate::system::commands::{ChannelCommand, OckamCommand};
 use failure::{Backtrace, Context, Fail};
 use ockam_kex::error::*;
-use ockam_system::commands::{ChannelCommand, OckamCommand};
 use ockam_vault::error::*;
 use std::{
     fmt,
