+++ implementations/rust/router/Cargo.toml
@@ -10,10 +10,7 @@ crate-type = ["rlib"]
 [profile.release]
 lto = true
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
 [dependencies]
-ockam-message = { version = "0.1", path = "../message" }
+ockam = { version = "0.1", path = "../ockam" }
 ockam-common = { version = "0.1", path = "../common" }
-ockam-system = { version = "0.1", path = "../system" }
