--- implementations/rust/system/Cargo.toml
@@ -1,18 +1,0 @@
-[package]
-name = "ockam-system"
-version = "0.1.0"
-authors = ["Ockam Developers"]
-edition = "2018"
-
-[lib]
-crate-type = ["rlib"]
-
-[profile.release]
-lto = true
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ockam-message = { version = "0.1", path = "../message" }
-ockam-vault = { version = "0.1", path = "../vault/traits" }
-
