+++ implementations/rust/transport/Cargo.toml
@@ -13,10 +13,9 @@ lto = true
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+ockam = { "version" = "0.1", path = "../ockam" }
 ockam-router = { version = "0.1", path = "../router" }
-ockam-message = { version = "0.1", path = "../message" }
 ockam-common = { version = "0.1", path = "../common" }
-ockam-system = { version = "0.1", path = "../system" }
 
 futures = "0.3"
 hashbrown = "0.9.1"
