+++ implementations/rust/worker/Cargo.toml
@@ -14,10 +14,9 @@ lto = true
 default = []
 
 [dependencies]
+ockam = { version = "0.1", path = "../ockam"}
 ockam-common = { version = "0.1", path = "../common" }
-ockam-message = { version = "0.1", path = "../message"}
 ockam-router = { version = "0.1", path = "../router"}
-ockam-system = { version = "0.1", path = "../system" }
 hashbrown = "0.9"
 
 [dev-dependencies]
